--- ui/src/features/project/pipelines/nodes/repo-node.module.less
@@ -1,60 +1,0 @@
-.node {
-  background-color: #eee;
-  border-radius: 10px;
-  display: flex;
-  flex-direction: column;
-  height: 100%;
-  border: 2px solid #eee;
-
-  h3 {
-    padding: 0.5em;
-    color: #333;
-    font-size: 15px;
-    font-weight: 600;
-    margin-left: 0.1em;
-    margin-bottom: 0;
-    z-index: 5;
-  }
-
-  .repoLabel {
-    color: gray;
-    font-size: 10px;
-    margin-bottom: 0.75em;
-    font-weight: 600;
-  }
-}
-
-.body {
-  background-color: white;
-  border-bottom-left-radius: 10px;
-  border-bottom-right-radius: 10px;
-  padding: 8px;
-  flex: 1;
-  display: flex;
-  @apply items-center justify-center;
-}
-
-.valueContainer {
-  width: 100%;
-  display: block;
-}
-
-.value {
-  display: block;
-  margin-top: 3px;
-  text-decoration: none;
-  white-space: nowrap;
-  overflow: hidden;
-  width: 100%;
-  max-width: 100%;
-  padding-bottom: 0.25em;
-  text-overflow: ellipsis;
-  color: black;
-  font-weight: 600;
-  font-size: 14px;
-  @apply font-mono items-center justify-center;
-
-  &:hover {
-    text-decoration: underline;
-  }
-}
