--- ui/src/features/project/pipelines/nodes/stage-node.module.less
@@ -1,55 +1,0 @@
-.node {
-  height: 100%;
-  border-radius: 10px;
-  transition: all 0.2s ease-in-out;
-  display: flex;
-  flex-direction: column;
-  border-width: 2px;
-  border-style: solid;
-  overflow: hidden;
-
-  h3 {
-    padding: 6px 9px;
-    color: white;
-    font-weight: 600;
-    font-size: 15px;
-    display: flex;
-    align-items: center;
-    margin-bottom: 0;
-  }
-
-  .body {
-    background-color: white;
-    text-align: center;
-    display: flex;
-    flex-direction: column;
-    flex: 1;
-    position: relative;
-  }
-
-  .smallLabel {
-    margin: 0;
-    margin-bottom: 0;
-    @apply text-gray-400;
-    text-transform: uppercase;
-    font-size: 9px;
-    font-weight: 500;
-  }
-}
-
-.faded {
-  background-color: #222 !important;
-  h3 {
-    color: #555;
-  }
-
-  .body {
-    background-color: #ccc;
-    color: gray;
-  }
-}
-
-.highlighted {
-  border-color: #1db6f9 !important;
-  box-shadow: 0 0 0 2px #1db6f9;
-}
